dnsdist: Use the shared library version of Quiche
Otherwise we have symbol conflicts with the symbols coming from
OpenSSL (DoT, DoH) and BoringSSL (DoQ, via Quiche). This does not
happen with a shared library because the symbols from BoringSSL
are in the same object than their user, Quiche, so they get picked
up first, and are not exported so they do not pollute dnsdist's
symbols namespace.